+++ cypress/e2e/header.cy.js
@@ -1,5 +1,4 @@
 import { getCurrentSession } from '@graasp/sdk';
-import { Platform } from '@graasp/ui';
 
 import { SIGN_IN_PATH } from '../../src/config/constants';
 import { HOME_PATH } from '../../src/config/paths';
@@ -21,9 +20,8 @@ describe('Header', () => {
     cy.setUpApi();
     cy.visit(HOME_PATH);
     // check navigation and display and interface doesn't crash
-    cy.get(
-      `#${APP_NAVIGATION_PLATFORM_SWITCH_BUTTON_IDS[Platform.Builder]}`,
-    ).click();
+    // todo: this is less robust than using the Platform contant from ui, but it was making cypress compile ui which is unnecessary.
+    cy.get(`#${APP_NAVIGATION_PLATFORM_SWITCH_BUTTON_IDS.Builder}`).click();
     cy.get(`#${APP_NAVIGATION_PLATFORM_SWITCH_ID}`).should('exist');
   });