+++ cypress/e2e/item/settings/itemSettings.cy.js
@@ -50,7 +50,7 @@ describe('Item Settings', () => {
             body: { settings },
           },
         }) => {
-          expect(settings.showChatbox).equals(false);
+          expect(settings?.showChatbox).equals(false);
           cy.wait(EDIT_ITEM_PAUSE);
           cy.get('@getItem').its('response.url').should('contain', itemId);
         },
@@ -73,7 +73,7 @@ describe('Item Settings', () => {
             body: { settings },
           },
         }) => {
-          expect(settings.showChatbox).equals(true);
+          expect(settings?.showChatbox).equals(true);
           cy.wait(EDIT_ITEM_PAUSE);
           cy.get('@getItem').its('response.url').should('contain', itemId);
         },