+++ cypress/support/createUtils.js
@@ -13,7 +13,8 @@ import { ITEM_TYPES } from '../fixtures/enums';
 
 // eslint-disable-next-line import/prefer-default-export
 export const createItem = (payload, options) => {
-  cy.get(`#${CREATE_ITEM_BUTTON_ID}`).click();
+  // click is forced to avoid errors with tooltips covering the button
+  cy.get(`#${CREATE_ITEM_BUTTON_ID}`).click({ force: true });
 
   switch (payload.type) {
     case ITEM_TYPES.LINK: