+++ src/components/item/ItemContent.tsx
@@ -5,6 +5,7 @@ import { UseQueryResult } from 'react-query';
 
 import { Api, MUTATION_KEYS } from '@graasp/query-client';
 import {
+  DEFAULT_LANG,
   DocumentItemExtraProperties,
   ItemType,
   PermissionLevel,
@@ -51,6 +52,7 @@ import { hooks, useMutation } from '../../config/queryClient';
 import {
   DOCUMENT_ITEM_TEXT_EDITOR_ID,
   ITEM_SCREEN_ERROR_ALERT_ID,
+  buildCancelButtonId,
   buildFileItemId,
   buildItemsTableId,
   buildSaveButtonId,
@@ -79,12 +81,16 @@ const FileContent = ({
   item,
   isEditing,
   onSaveCaption,
+  onCancelCaption,
   saveButtonId,
+  cancelButtonId,
 }: {
   item: LocalFileItemTypeRecord | S3FileItemTypeRecord;
   isEditing: boolean;
   onSaveCaption: (text: string) => void;
+  onCancelCaption: (text: string) => void;
   saveButtonId: string;
+  cancelButtonId: string;
 }): JSX.Element => {
   const {
     data: fileContent,
@@ -113,8 +119,10 @@ const FileContent = ({
         id={buildFileItemId(item.id)}
         item={item}
         onSaveCaption={onSaveCaption}
+        onCancelCaption={onCancelCaption}
         pdfViewerLink={buildPdfViewerLink(GRAASP_ASSETS_URL)}
         saveButtonId={saveButtonId}
+        cancelButtonId={cancelButtonId}
       />
     </StyledContainer>
   );
@@ -128,13 +136,17 @@ const LinkContent = ({
   member,
   isEditing,
   onSaveCaption,
+  onCancelCaption,
   saveButtonId,
+  cancelButtonId,
 }: {
   item: EmbeddedLinkItemTypeRecord;
   member: MemberRecord;
   isEditing: boolean;
   onSaveCaption: (caption: string) => void;
+  onCancelCaption: (caption: string) => void;
   saveButtonId: string;
+  cancelButtonId: string;
 }): JSX.Element => (
   <StyledContainer>
     <LinkItem
@@ -143,7 +155,9 @@ const LinkContent = ({
       item={item}
       editCaption={isEditing}
       onSaveCaption={onSaveCaption}
+      onCancelCaption={onCancelCaption}
       saveButtonId={saveButtonId}
+      cancelButtonId={cancelButtonId}
       height={ITEM_DEFAULT_HEIGHT}
       showButton={Boolean(
         item.settings?.showLinkButton ?? DEFAULT_LINK_SHOW_BUTTON,
@@ -164,12 +178,14 @@ const DocumentContent = ({
   onSaveDocument,
   onCancel,
   saveButtonId,
+  cancelButtonId,
 }: {
   item: DocumentItemTypeRecord;
   isEditing: boolean;
   onSaveDocument: (update: DocumentItemExtraProperties) => void;
   onCancel: () => void;
   saveButtonId: string;
+  cancelButtonId: string;
 }): JSX.Element => {
   const { t: translateCommon } = useCommonTranslation();
 
@@ -199,8 +215,14 @@ const DocumentContent = ({
             onContentSave={onSave}
             onFlavorChange={(f) => setFlavor(f || undefined)}
             saveButtonId={saveButtonId}
+            cancelButtonId={cancelButtonId}
           />
-          <Button onClick={onCancel} variant="text" sx={{ m: 1 }}>
+          <Button
+            id={cancelButtonId}
+            onClick={onCancel}
+            variant="text"
+            sx={{ m: 1 }}
+          >
             {translateCommon(COMMON.CANCEL_BUTTON)}
           </Button>
           <SaveButton
@@ -231,27 +253,44 @@ const AppContent = ({
   permission,
   isEditing,
   onSaveCaption,
+  onCancelCaption,
   saveButtonId,
+  cancelButtonId,
 }: {
   item: AppItemTypeRecord;
   member: MemberRecord;
   permission: PermissionLevel;
   isEditing: boolean;
   onSaveCaption: (caption: string) => void;
+  onCancelCaption: (caption: string) => void;
   saveButtonId: string;
+  cancelButtonId: string;
 }): JSX.Element => (
   <AppItem
     isResizable
     item={item}
-    apiHost={API_HOST}
     editCaption={isEditing}
     onSaveCaption={onSaveCaption}
+    onCancelCaption={onCancelCaption}
     saveButtonId={saveButtonId}
-    member={member}
+    cancelButtonId={cancelButtonId}
     height={ITEM_DEFAULT_HEIGHT}
-    permission={permission}
-    requestApiAccessToken={Api.requestApiAccessToken}
-    context={CONTEXT_BUILDER}
+    requestApiAccessToken={(payload) =>
+      Api.requestApiAccessToken(payload, { API_HOST })
+    }
+    contextPayload={{
+      apiHost: API_HOST,
+      itemId: item.id,
+      memberId: member.id,
+      permission,
+      settings: item.settings,
+      lang:
+        // todo: remove once it is added in ItemSettings type in sdk
+        (item.settings?.lang as string | undefined) ||
+        member.extra?.lang ||
+        DEFAULT_LANG,
+      context: CONTEXT_BUILDER,
+    }}
   />
 );
 
@@ -312,6 +351,7 @@ const H5PContent = ({ item }: { item: H5PItemTypeRecord }): JSX.Element => {
   return (
     <H5PItem
       itemId={item.id}
+      itemName={item.name}
       contentId={contentId}
       integrationUrl={H5P_INTEGRATION_URL}
     />
@@ -394,6 +434,10 @@ const ItemContent = ({
     setEditingItemId(null);
   };
 
+  const onCancelCaption = (_caption: string) => {
+    setEditingItemId(null);
+  };
+
   const onSaveDocument = ({ content, flavor }: DocumentItemExtraProperties) => {
     editItem({
       id: item.id,
@@ -407,6 +451,7 @@ const ItemContent = ({
   };
 
   const saveButtonId = buildSaveButtonId(item.id);
+  const cancelButtonId = buildCancelButtonId(item.id);
 
   switch (item.type) {
     case ItemType.LOCAL_FILE:
@@ -416,7 +461,9 @@ const ItemContent = ({
           item={item}
           isEditing={isEditing}
           onSaveCaption={onSaveCaption}
+          onCancelCaption={onCancelCaption}
           saveButtonId={saveButtonId}
+          cancelButtonId={cancelButtonId}
         />
       );
     }
@@ -427,7 +474,9 @@ const ItemContent = ({
           member={member}
           isEditing={isEditing}
           onSaveCaption={onSaveCaption}
+          onCancelCaption={onCancelCaption}
           saveButtonId={saveButtonId}
+          cancelButtonId={cancelButtonId}
         />
       );
     case ItemType.DOCUMENT:
@@ -438,6 +487,7 @@ const ItemContent = ({
           onSaveDocument={onSaveDocument}
           onCancel={onCancel}
           saveButtonId={saveButtonId}
+          cancelButtonId={cancelButtonId}
         />
       );
     case ItemType.APP:
@@ -447,7 +497,9 @@ const ItemContent = ({
           member={member}
           isEditing={isEditing}
           onSaveCaption={onSaveCaption}
+          onCancelCaption={onCancelCaption}
           saveButtonId={saveButtonId}
+          cancelButtonId={cancelButtonId}
           permission={permission}
         />
       );