+++ src/components/main/Items.tsx
@@ -7,10 +7,11 @@ import { hooks } from '../../config/queryClient';
 import { ITEM_LAYOUT_MODES } from '../../enums';
 import { useLayoutContext } from '../context/LayoutContext';
 import { useItemSearch } from '../item/ItemSearch';
+import { useItemsStatuses } from '../table/BadgesCellRenderer';
 import ItemsGrid from './ItemsGrid';
 import ItemsTable from './ItemsTable';
 
-const { useManyItemMemberships } = hooks;
+const { useManyItemMemberships, useTags, useItemsTags } = hooks;
 
 type Props = {
   id: string;
@@ -50,18 +51,29 @@ const Items = ({
 }: Props): JSX.Element => {
   const { mode } = useLayoutContext();
   const itemSearch = useItemSearch(items);
+  const itemsToDisplay = itemSearch.results;
   const { data: manyMemberships, isLoading: isMembershipsLoading } =
     useManyItemMemberships(
       enableMemberships
-        ? itemSearch?.results?.map(({ id: itemId }) => itemId).toArray()
+        ? itemsToDisplay?.map(({ id: itemId }) => itemId).toArray()
         : [],
     );
+  const { data: tagList, isLoading: isLoadingTagList } = useTags();
+  const { data: itemsTags } = useItemsTags(
+    itemsToDisplay.map((r) => r.id).toJS(),
+  );
+  const itemsStatuses = useItemsStatuses({
+    items: itemsToDisplay,
+    itemsTags,
+    tagList,
+  });
+
   // todo: disable depending on showCreator
   const { data: creators } = hooks.useMembers(
     Object.keys(items?.groupBy(({ creator }) => creator)?.toJS() ?? []),
   );
 
-  if (isMembershipsLoading) {
+  if (isMembershipsLoading || isLoadingTagList) {
     return <Loader />;
   }
 
@@ -71,8 +83,9 @@ const Items = ({
         <ItemsGrid
           parentId={parentId}
           title={title}
-          items={itemSearch.results}
+          items={itemsToDisplay}
           manyMemberships={manyMemberships}
+          itemsStatuses={itemsStatuses}
           // This enables the possiblity to display messages (item is empty, no search result)
           itemSearch={itemSearch}
           headerElements={[itemSearch.input, ...headerElements]}
@@ -87,8 +100,9 @@ const Items = ({
           actions={actions}
           tableTitle={title}
           defaultSortedColumn={defaultSortedColumn}
-          items={itemSearch.results}
+          items={itemsToDisplay}
           manyMemberships={manyMemberships}
+          itemsStatuses={itemsStatuses}
           headerElements={[itemSearch.input, ...headerElements]}
           isSearching={Boolean(itemSearch.text)}
           ToolbarActions={ToolbarActions}