+++ src/components/main/ItemsTable.tsx
@@ -33,6 +33,7 @@ import { formatDate } from '../../utils/date';
 import { useCurrentUserContext } from '../context/CurrentUserContext';
 import FolderDescription from '../item/FolderDescription';
 import ActionsCellRenderer from '../table/ActionsCellRenderer';
+import BadgesCellRenderer, { ItemsStatuses } from '../table/BadgesCellRenderer';
 import NameCellRenderer from '../table/ItemNameCellRenderer';
 import MemberNameCellRenderer from '../table/MemberNameCellRenderer';
 import ItemsToolbar from './ItemsToolbar';
@@ -42,7 +43,8 @@ const { useItem } = hooks;
 type Props = {
   id?: string;
   items: List<ItemRecord>;
-  manyMemberships: List<List<ItemMembershipRecord>>;
+  manyMemberships?: List<List<ItemMembershipRecord>>;
+  itemsStatuses?: ItemsStatuses;
   tableTitle: string;
   headerElements?: JSX.Element[];
   isSearching?: boolean;
@@ -66,6 +68,7 @@ const ItemsTable: FC<Props> = ({
   id: tableId = '',
   items: rows = List(),
   manyMemberships = List(),
+  itemsStatuses,
   headerElements = [],
   isSearching = false,
   actions,
@@ -84,6 +87,10 @@ const ItemsTable: FC<Props> = ({
   const { itemId } = useParams();
   const { data: parentItem } = useItem(itemId);
   const { data: member } = useCurrentUserContext();
+
+  const noStatusesToShow = !Object.values(itemsStatuses)
+    .map((obj) => Object.values(obj).some((e) => e === true))
+    .some((e) => e === true);
 
   const mutation = useMutation<
     unknown,
@@ -167,19 +174,39 @@ const ItemsTable: FC<Props> = ({
     member,
   });
 
+  const BadgesComponent = BadgesCellRenderer({
+    itemsStatuses,
+  });
+
   // never changes, so we can use useMemo
   const columnDefs = useMemo(() => {
     const columns: ColDef[] = [
       {
+        field: 'name',
+        headerName: translateBuilder(BUILDER.ITEMS_TABLE_NAME_HEADER),
         headerCheckboxSelection: true,
         checkboxSelection: true,
-        headerName: translateBuilder(BUILDER.ITEMS_TABLE_NAME_HEADER),
         cellRenderer: NameCellRenderer(showThumbnails),
         flex: 4,
         comparator: GraaspTable.textComparator,
         sort: defaultSortedColumn?.name,
-        field: 'name',
         tooltipField: 'name',
+      },
+      {
+        field: 'status',
+        headerName: translateBuilder(BUILDER.ITEMS_TABLE_STATUS_HEADER),
+        cellRenderer: BadgesComponent,
+        hide: noStatusesToShow,
+        type: 'rightAligned',
+        flex: 1,
+        suppressAutoSize: true,
+        maxWidth: 100,
+        cellStyle: {
+          display: 'flex',
+          flexDirection: 'row',
+          justifyContent: 'center',
+          alignItems: 'center',
+        },
       },
       {
         field: 'type',
@@ -187,14 +214,16 @@ const ItemsTable: FC<Props> = ({
         type: 'rightAligned',
         cellRenderer: ({ data }: { data: DiscriminatedItem }) =>
           translateEnums(data.type),
-        flex: 2,
+        minWidth: 90,
+        maxWidth: 120,
         comparator: GraaspTable.textComparator,
         sort: defaultSortedColumn?.type,
       },
       {
         field: 'updatedAt',
         headerName: translateBuilder(BUILDER.ITEMS_TABLE_UPDATED_AT_HEADER),
-        flex: 2,
+        maxWidth: 160,
+        minWidth: 80,
         type: 'rightAligned',
         valueFormatter: dateColumnFormatter,
         comparator: GraaspTable.dateComparator,
@@ -207,22 +236,21 @@ const ItemsTable: FC<Props> = ({
         headerName: translateBuilder(BUILDER.ITEMS_TABLE_ACTIONS_HEADER),
         colId: 'actions',
         type: 'rightAligned',
-        flex: 3,
         cellStyle: {
           paddingLeft: '0!important',
           paddingRight: '0!important',
           textAlign: 'right',
         },
         sortable: false,
+        suppressAutoSize: true,
         // prevent ellipsis for small screens
-        minWidth: 165,
+        minWidth: 140,
       },
     ];
 
     if (showCreator) {
-      columns.splice(1, 0, {
+      columns.splice(2, 0, {
         field: 'creator',
-        flex: 3,
         headerName: translateBuilder(BUILDER.ITEMS_TABLE_CREATOR_HEADER),
         colId: 'creator',
         type: 'rightAligned',
@@ -245,6 +273,7 @@ const ItemsTable: FC<Props> = ({
     translateBuilder,
     defaultSortedColumn,
     ActionComponent,
+    BadgesComponent,
     actions,
     showThumbnails,
   ]);